Moved some utils into backend/utils.py
