--- backend/auth/utils.py
@@ -1,43 +1,0 @@
-from uuid import UUID
-from rest_framework_simplejwt.authentication import JWTAuthentication
-import re
-JWT_authenticator = JWTAuthentication()
-
-def isUUID(val):
-    try:
-        UUID(val)
-        return True
-    except:
-        return False
-
-def isAuthorized(request, author_uuid):
-    ''' Checks if the requester is authorized to do whatever method that's requested.
-        author_uuid = uuid of the author in the requested url
-    '''
-    res = JWT_authenticator.authenticate(request)
-    if res is not None:
-        user, token = res;
-        requesterUUID = user.uuid
-        
-        # if the requester is not what they say they are (aren't the actual author)
-        if str(requesterUUID) != author_uuid:
-            return False
-        return True
-    else:
-        return False
-
-def check_github_valid(request):
-    '''Checks if the given github (if exists) is a valid github link'''
-    github = request.data.get("github")
-
-    # Check that github is not null and is not an empty string
-    if github is not None and github != '':
-        # check if the github string given starts with http
-        if github.startswith('http') and github.split("/")[2] == 'github.com':
-            return True
-        else:
-            return False
-    else:
-        # return True since we allow a users to not have to pull
-        # in activity from their github (blank URL)
-        return True
