+++ backend/like/views.py
@@ -11,8 +11,7 @@ from like.models import Like
 from comments.models import Comment
 from django.db.models import Q
 from like.serializers import LikeCommentSerializer, LikePostSerializer, LikeAuthorSerializer
-# Create your views here.
-from auth.utils import isUUID, isAuthorized
+from backend.utils import isUUID, isAuthorized
 
 class LikesPostApiView(GenericAPIView):
     """
